change: Split RoleCommand and RolesCommand into two modules
