+++ src/gmp/gmp.ts
@@ -44,7 +44,8 @@ import LoginCommand from 'gmp/commands/login';
 import PerformanceCommand from 'gmp/commands/performance';
 import {PortListCommand, PortListsCommand} from 'gmp/commands/portlists';
 import {ReportCommand, ReportsCommand} from 'gmp/commands/reports';
-import {RoleCommand, RolesCommand} from 'gmp/commands/roles';
+import RoleCommand from 'gmp/commands/role';
+import RolesCommand from 'gmp/commands/roles';
 import {TaskCommand, TasksCommand} from 'gmp/commands/tasks';
 import TrashCanCommand from 'gmp/commands/trashcan';
 import UserCommand from 'gmp/commands/user';
